+++ src/components/script/ScriptEditorEvent.tsx
@@ -1,4 +1,10 @@
-import React, { useCallback, useRef, useState } from "react";
+import React, {
+  useCallback,
+  useContext,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
 import {
   DragSourceMonitor,
   DropTargetMonitor,
@@ -8,6 +14,7 @@ import {
 import entitiesActions from "store/features/entities/entitiesActions";
 import { scriptEventSelectors } from "store/features/entities/entitiesState";
 import editorActions from "store/features/editor/editorActions";
+import settingsActions from "store/features/settings/settingsActions";
 import {
   ScriptEventParentType,
   ScriptEventsRef,
@@ -25,15 +32,21 @@ import {
   ScriptEventRenameInputCompleteButton,
   ScriptEditorChildrenWrapper,
   ScriptEditorChildrenLabel,
+  ScriptEventHeaderBreakpointIndicator,
 } from "ui/scripting/ScriptEvents";
-import { ArrowIcon, CheckIcon, CommentIcon } from "ui/icons/Icons";
+import {
+  ArrowIcon,
+  BreakpointIcon,
+  CheckIcon,
+  CommentIcon,
+} from "ui/icons/Icons";
 import { FixedSpacer } from "ui/spacing/Spacing";
 import ScriptEventForm from "./ScriptEventForm";
 import l10n from "shared/lib/lang/l10n";
 import { ScriptEditorEventHelper } from "./ScriptEditorEventHelper";
 import ItemTypes from "renderer/lib/dnd/itemTypes";
 import { DropdownButton } from "ui/buttons/DropdownButton";
-import { MenuDivider, MenuItem, MenuOverlay } from "ui/menu/Menu";
+import { MenuDivider, MenuItem, MenuItemIcon, MenuOverlay } from "ui/menu/Menu";
 import clipboardActions from "store/features/clipboard/clipboardActions";
 import { ClipboardTypeScriptEvents } from "store/features/clipboard/clipboardTypes";
 import { RelativePortal } from "ui/layout/RelativePortal";
@@ -45,6 +58,8 @@ import { ScriptEventSymbolEditorWrapper } from "components/forms/symbols/SymbolE
 import { EVENT_CALL_CUSTOM_EVENT, EVENT_COMMENT, EVENT_END } from "consts";
 import { selectScriptEventDefs } from "store/features/scriptEventDefs/scriptEventDefsState";
 import { useAppDispatch, useAppSelector } from "store/hooks";
+import { ScriptEditorContext } from "components/script/ScriptEditorContext";
+import { getSettings } from "store/features/settings/settingsState";
 
 interface ScriptEditorEventProps {
   id: string;
@@ -67,6 +82,7 @@ const ScriptEditorEvent = React.memo(
     nestLevel = 0,
   }: ScriptEditorEventProps) => {
     const dispatch = useAppDispatch();
+    const context = useContext(ScriptEditorContext);
     const dragRef = useRef<HTMLDivElement>(null);
     const dropRef = useRef<HTMLDivElement>(null);
     const headerRef = useRef<HTMLDivElement>(null);
@@ -85,6 +101,12 @@ const ScriptEditorEvent = React.memo(
     );
     const scriptEventDefs = useAppSelector((state) =>
       selectScriptEventDefs(state)
+    );
+    const breakpointEnabled = useAppSelector(
+      (state) =>
+        getSettings(state).debuggerBreakpoints.findIndex(
+          (b) => b.scriptEventId === id
+        ) > -1
     );
 
     const onFetchClipboard = useCallback(() => {
@@ -169,6 +191,15 @@ const ScriptEditorEvent = React.memo(
         entitiesActions.toggleScriptEventDisableElse({ scriptEventId: id })
       );
     }, [dispatch, id]);
+
+    const toggleBreakpoint = useCallback(() => {
+      dispatch(
+        settingsActions.toggleBreakpoint({
+          scriptEventId: id,
+          context,
+        })
+      );
+    }, [context, dispatch, id]);
 
     const editCustomEvent = useCallback(() => {
       const customEventId = scriptEvent?.args?.customEventId;
@@ -316,6 +347,14 @@ const ScriptEditorEvent = React.memo(
       setAddOpen(false);
     }, []);
 
+    const isExecuting = scriptEvent?.id === context.executingId;
+
+    useEffect(() => {
+      if (isExecuting && headerRef.current) {
+        headerRef.current.scrollIntoView();
+      }
+    }, [isExecuting]);
+
     if (!scriptEvent) {
       return null;
     }
@@ -365,6 +404,7 @@ const ScriptEditorEvent = React.memo(
               comment={Boolean(commented || isComment)}
               nestLevel={nestLevel}
               altBg={index % 2 === 0}
+              isExecuting={isExecuting}
             >
               {isVisible && (
                 <>
@@ -406,6 +446,13 @@ const ScriptEditorEvent = React.memo(
                       />
                     )}
                   </ScriptEventHeaderTitle>
+                  {breakpointEnabled && (
+                    <ScriptEventHeaderBreakpointIndicator
+                      title={l10n("FIELD_BREAKPOINT")}
+                    >
+                      <BreakpointIcon />
+                    </ScriptEventHeaderBreakpointIndicator>
+                  )}
 
                   <DropdownButton
                     size="small"
@@ -439,6 +486,14 @@ const ScriptEditorEvent = React.memo(
                           : l10n("MENU_DISABLE_ELSE")}
                       </MenuItem>
                     )}
+                    <MenuItem onClick={toggleBreakpoint}>
+                      <span>{l10n("MENU_SET_BREAKPOINT")}</span>
+                      {breakpointEnabled && (
+                        <MenuItemIcon>
+                          <CheckIcon />
+                        </MenuItemIcon>
+                      )}
+                    </MenuItem>
                     <MenuDivider />
                     <MenuItem onClick={() => onOpenAddMenu(true)}>
                       {l10n("MENU_INSERT_EVENT_BEFORE")}