+++ src/components/script/ScriptEventFormMatharea.tsx
@@ -35,7 +35,7 @@ const ScriptEventFormMathArea: FC<ScriptEventFormMathAreaProps> = ({
 
   useEffect(() => {
     setVariables(
-      namedVariablesByContext(context, entityId, variablesLookup, customEvent)
+      namedVariablesByContext(context, variablesLookup, customEvent)
     );
   }, [entityId, variablesLookup, context, customEvent]);