+++ src/consts.ts
@@ -192,6 +192,14 @@ export const defaultProjectSettings: SettingsState = {
   customControlsB: ["Control", "k", "x"],
   customControlsStart: ["Enter"],
   customControlsSelect: ["Shift"],
+  debuggerEnabled: false,
+  debuggerScriptType: "editor",
+  debuggerVariablesFilter: "all",
+  debuggerCollapsedPanes: [],
+  debuggerPauseOnScriptChanged: false,
+  debuggerPauseOnWatchedVariableChanged: false,
+  debuggerBreakpoints: [],
+  debuggerWatchedVariables: [],
 };
 
 export {