+++ src/menu.ts
@@ -192,7 +192,14 @@ const buildMenu = async (plugins: MenuItemConstructorOptions[] = []) => {
           label: l10n("MENU_RUN"),
           accelerator: "CommandOrControl+B",
           click: () => {
-            notifyListeners("run");
+            notifyListeners("run", false);
+          },
+        },
+        {
+          label: l10n("MENU_RUN_WITH_DEBUGGING"),
+          accelerator: "CommandOrControl+Alt+B",
+          click: () => {
+            notifyListeners("run", true);
           },
         },
         {