+++ src/renderer/lib/api/setup.ts
@@ -32,8 +32,10 @@ import type { FontAssetData } from "lib/project/loadFontData";
 import type { AvatarAssetData } from "lib/project/loadAvatarData";
 import type { EmoteAssetData } from "lib/project/loadEmoteData";
 import type { NavigationSection } from "store/features/navigation/navigationState";
-import { ScriptEventDefs } from "shared/lib/scripts/scriptDefHelpers";
+import type { ScriptEventDefs } from "shared/lib/scripts/scriptDefHelpers";
 import type { MenuZoomType } from "menu";
+import type { DebuggerDataPacket } from "shared/lib/debugger/types";
+import type { SceneMapData, VariableMapData } from "lib/compiler/compileData";
 
 interface L10NLookup {
   [key: string]: string | boolean | undefined;
@@ -44,6 +46,7 @@ export type BuildOptions = {
   profile: boolean;
   engineFields: EngineFieldSchema[];
   exportBuild: boolean;
+  debugEnabled?: boolean;
 };
 
 const createSubscribeAPI = <
@@ -275,6 +278,24 @@ const APISetup = {
     writeBuffer: (format: string, buffer: Buffer): Promise<void> =>
       ipcRenderer.invoke("clipboard:write-buffer", format, buffer),
   },
+  debugger: {
+    pause: () => ipcRenderer.invoke("debugger:pause"),
+    resume: () => ipcRenderer.invoke("debugger:resume"),
+    setPauseOnScriptChanged: (enabled: boolean) =>
+      ipcRenderer.invoke("debugger:pause-on-script", enabled),
+    setPauseOnWatchVariableChanged: (enabled: boolean) =>
+      ipcRenderer.invoke("debugger:pause-on-var", enabled),
+    setGlobal: (symbol: string, value: number) =>
+      ipcRenderer.invoke("debugger:set-global", symbol, value),
+    step: () => ipcRenderer.invoke("debugger:step"),
+    stepFrame: () => ipcRenderer.invoke("debugger:step-frame"),
+    setBreakpoints: (breakpoints: string[]) =>
+      ipcRenderer.invoke("debugger:set-breakpoints", breakpoints),
+    setWatchedVariableIds: (variableIds: string[]) =>
+      ipcRenderer.invoke("debugger:set-watched", variableIds),
+    sendToProjectWindow: (data: DebuggerDataPacket) =>
+      ipcRenderer.send("debugger:data-receive", data),
+  },
   events: {
     menu: {
       saveProject:
@@ -303,7 +324,9 @@ const APISetup = {
       zoom: createSubscribeAPI<
         (event: IpcRendererEvent, zoom: MenuZoomType) => void
       >("menu:zoom"),
-      run: createSubscribeAPI<(event: IpcRendererEvent) => void>("menu:run"),
+      run: createSubscribeAPI<
+        (event: IpcRendererEvent, debugEnabled: boolean) => void
+      >("menu:run"),
       build:
         createSubscribeAPI<
           (event: IpcRendererEvent, buildType: BuildType) => void
@@ -329,6 +352,25 @@ const APISetup = {
       data: createSubscribeAPI<
         (event: IpcRendererEvent, data: MusicDataPacket) => void
       >("music:data"),
+    },
+    debugger: {
+      data: createSubscribeAPI<
+        (event: IpcRendererEvent, data: DebuggerDataPacket) => void
+      >("debugger:data"),
+      symbols:
+        createSubscribeAPI<
+          (
+            event: IpcRendererEvent,
+            data: {
+              variableMap: Record<string, VariableMapData>;
+              sceneMap: Record<string, SceneMapData>;
+              gbvmScripts: Record<string, string>;
+            }
+          ) => void
+        >("debugger:symbols"),
+      disconnected: createSubscribeAPI<(event: IpcRendererEvent) => void>(
+        "debugger:disconnected"
+      ),
     },
     settings: {
       uiScaleChanged: createSubscribeAPI<